fix(openvino): rate-limit gemma4 step() WARN via shared limiter (#30)
- gemma4's step() had the byte-identical unlimited per-call WARN that
  runtime.rs fixed: a dropped upstream → recv fast-errors → relay loop
  floods at ~90k/s. Same latent twin of the 770 GB chain.log bug.
- lift StepWarnLimiter to crate-shared warn_limit module; reuse in both
  runtime and gemma4 (first-failure WARN + 30s summary + recovery info)
- demote gemma4 step_first inner WARN to debug (outer step() limits it)
- dist_spec worker assessed + left: its CPU-speed path (socket-closed) is
  already throttled; else-branch errors are recv-bounded, not a flood
